Unlock and Open functionality
I have integrated what I think will be able to use the logic to make sure you cannot open doors when you dont have the correct key or you dont have the key. And you cant open a door when you arent in position, or when the door isnt unlocked!