Restart and Discard buttons
Restart operating at full functionality. Discard is not entirely working however just yet. As there is an inability to pick up anything from the floor as the positions of objects have not all been initialised and the player is unable to navigate to spaces where the objects which are specified are.